Extract enqueueInRustyHook and metric to their own files
